+++ CHANGES.md
@@ -14,6 +14,7 @@ Changes in 4.5
 - Add Report Cards header action hook in ReportCards.php
 - Add Report Cards PDF header action hook in ReportCards.fnc.php
 - HTML add autocomplete="off" to select inputs in Side.php
+- Add GetFullYearMP() function in GetMP.php
 
 Changes in 4.4
 --------------